push

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -7,7 +7,7 @@
 */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int unmatch = n ^ m, set = 0;
+	unsigned long int unmatch = n ^ m, set = 0;
 
 	while (unmatch)
 	{
